test: challenge 서비스 계층 테스트 추가 작성
